requirements.txt
python-telegram-bot==20.5
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.0